requirements-frozen.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
cryptography==45.0.4
cssselect2==0.8.0
diff-match-patch==20241021
Django==5.2.3
django-bootstrap-forms==0.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-forms-bootstrap==3.1.0
django-humanize==0.1.2
django-import-export==4.3.8
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
Faker==37.8.0
fonttools==4.59.1
gunicorn==23.0.0
html5lib==1.1
humanize==4.12.3
idna==3.10
import-export==0.3.1
lxml==6.0.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
polib==1.2.0
psycopg==3.2.9
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.6.1
pyphen==0.17.2
python-barcode==0.15.1
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
SQLite4==0.1.1
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==58.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
